-- added pages to manage worker instances -- instances are tagged by their role (sql server or worker), making it easier to filter them -- moved some of the load balancer and worker creation code out of 'start' into the loadbalancer and worker files

+++ manager/stop.py
@@ -13,6 +13,10 @@ def stop():
     elb.delete_load_balancer(LoadBalancerName=loadbalancer.elb_name)
     all_ec2_instances = ec2.instances.all()
     for instance in all_ec2_instances:
-        instance.terminate()
+        if instance.tags is not None:
+            for tag in instance.tags:
+                if tag['Key'] == 'Role' and tag['Value'] == 'worker':
+                    instance.terminate()
+                    break
 
     return redirect(url_for('index'))
